requirements.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.7
drf-spectacular==0.26.5
gunicorn==21.2.0
dj-database-url==2.1.0